+++ dqops/src/main/java/com/dqops/utils/version/CopyVersionToFilesPostProcessor.java
@@ -1,17 +1,11 @@
 /*
- * Copyright © 2021 DQOps (support@dqops.com)
+ * Copyright © 2021-Present DQOps, Documati sp. z o.o. (support@dqops.com)
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * This file is licensed under the Business Source License 1.1,
+ * which can be found in the root directory of this repository.
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Change Date: This file will be licensed under the Apache License, Version 2.0,
+ * four (4) years from its last modification date.
  */
 package com.dqops.utils.version;
 
@@ -68,6 +62,10 @@ public class CopyVersionToFilesPostProcessor {
 
         updateVersionInBannerTxt(repositoryRootPath.resolve("dqops/src/main/resources/banner.txt"), version);
         updateVersionInIntelliJRunConfig(repositoryRootPath.resolve(".run/dqo run.run.xml"), version);
+
+        updateVersionInLicense(repositoryRootPath.resolve("LICENSE"), version);
+        updateVersionInLicense(repositoryRootPath.resolve("home/licenses/BUSL_v_1_1/license.txt"), version);
+        updateVersionInLicense(repositoryRootPath.resolve("distribution/python/LICENSE"), version);
     }
 
     /**
@@ -265,6 +263,28 @@ public class CopyVersionToFilesPostProcessor {
 
         if (modified) {
             Files.write(pathToVersionPy, lines, StandardCharsets.UTF_8);
+        }
+    }
+
+    /**
+     * Replace the version number in the LICENSE file
+     * @param pathToDqoScript Path to a LICENSE
+     * @param version New version number.
+     * @throws Exception
+     */
+    public static void updateVersionInLicense(Path pathToDqoScript, String version) throws Exception {
+        List<String> lines = Files.readAllLines(pathToDqoScript, StandardCharsets.UTF_8);
+        for (int i = 0; i < lines.size(); i++) {
+            String line = lines.get(i);
+            if (line.startsWith("Licensed Work: DQOps Data Quality Operations Center ")) {
+                String correctLine = "Licensed Work: DQOps Data Quality Operations Center " + version;
+                if (!Objects.equals(line, correctLine)) {
+                    lines.set(i, correctLine);
+                    Files.write(pathToDqoScript, lines, StandardCharsets.UTF_8);
+                }
+
+                return;
+            }
         }
     }
 }