+++ dqops/src/test/java/com/dqops/connectors/hana/HanaContainer.java
@@ -1,4 +1,14 @@
 /*
+ * Copyright © 2021-Present DQOps, Documati sp. z o.o. (support@dqops.com)
+ *
+ * This file is licensed under the Business Source License 1.1,
+ * which can be found in the root directory of this repository.
+ *
+ * Change Date: This file will be licensed under the Apache License, Version 2.0,
+ * four (4) years from its last modification date.
+ */
+
+/*
  * This code is intended for the internal use by the DQOps team for the development purpose only.
  * Using SAP HANA Express requires acceptance of the SAP license.
  */