+++ home/rules/percentile/anomaly_differencing_percentile_moving_average_30_days.py
@@ -1,18 +1,10 @@
+#  Copyright © 2021-Present DQOps, Documati sp. z o.o. (support@dqops.com)
 #
-# Copyright © 2024 DQOps (support@dqops.com)
+#  This file is licensed under the Business Source License 1.1,
+#  which can be found in the root directory of this repository.
 #
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
+#  Change Date: This file will be licensed under the Apache License, Version 2.0,
+#  four (4) years from its last modification date.
 
 from datetime import datetime
 from typing import Sequence, Dict