+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Fixes for working on a machine with a low CPU count (better MaxDOP management for Python processes)
 * Current DQ status calculated across different table comparisons.
+* Compare the first 1000 records, dropping the rest