+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 * Fixes for working on a machine with a low CPU count (better MaxDOP management for Python processes)
 * Current DQ status calculated across different table comparisons.
 * Compare the first 1000 records, dropping the rest
-* Apply closed source code (2025-07-05 21:04)
+* Apply closed source code (2025-07-14 23:34)