+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * Current DQ status calculated across different table comparisons.
 * Compare the first 1000 records, dropping the rest
 * Apply closed source code (2025-07-14 23:34)
+* Change the license to BUSL 1.1